refactor: split semantic and LLVM codegen by AST responsibility
Break the semantic analyzer and the LLVM backend into smaller modules organized around clear responsibilities: public facade modules, statement handling, expression-specific handlers, and shared helper code. This keeps the compiler-facing API unchanged while making the internal layout easier to navigate and far less conflict-prone for parallel work.

The semantic module now exposes a dedicated README that documents the folder layout, the analyzer flow, ownership boundaries, and the rules for placing code in analyzer, expr, statement, or helper modules. The codegen README was also updated so the LLVM backend documentation matches the current structure.

When adding a new expression in the future, follow this path: extend the AST and parser first, add a dedicated semantic handler file under src/semantic/expr and wire it from expr/mod.rs, place only reusable cross-node utilities under src/semantic/helper, add the matching LLVM emission file under src/codegen/llvm/expr and wire it from expr/mod.rs there as well, and finish by adding semantic and codegen tests for the new node. Keep node-specific logic out of helper/ so the helper layer does not become a new monolith.
