+++ src/semantic/expr/mod.rs
@@ -1,0 +1,29 @@
+mod binary;
+mod block;
+mod builtin_call;
+mod destructive_assign;
+mod let_in;
+mod literal;
+mod unary;
+mod variable;
+
+use crate::parser::expression::Expr;
+
+use super::{SemanticType, analyzer::SemanticAnalyzer};
+
+impl SemanticAnalyzer {
+    pub(super) fn check_expr(&mut self, expr: &Expr, source: &str) -> Option<SemanticType> {
+        match expr {
+            Expr::Literal { value, .. } => Some(self.check_literal(value)),
+            Expr::DestructiveAssign(assign) => self.check_destructive_assign(assign, source),
+            Expr::Variable { name, span } => self.check_variable(name, *span, source),
+            Expr::Unary(unary) => self.check_unary_expr(unary, source),
+            Expr::Block(block) => self.check_block_expr(block, source),
+            Expr::LetIn(let_in) => self.check_let_in_expr(let_in, source),
+            Expr::BuiltinCall(call) => {
+                self.check_builtin_call(call.function, &call.args, call.span, source)
+            }
+            Expr::Binary(binary) => self.check_binary_expr(binary, source),
+        }
+    }
+}
