+++ src/semantic/mod.rs
@@ -1,597 +1,9 @@
-use std::collections::HashMap;
+mod analyzer;
+mod expr;
+mod helper;
+mod statement;
 #[cfg(test)]
 mod tests;
 
-use crate::{
-    error::{CompilerError, ErrorCategory, offset_to_line_column},
-    parser::expression::{
-        BinaryOp, BlockExpr, BuiltinFunction, DestructiveAssignExpr, Expr, LetInExpr, Literal,
-        Program, Span, Statement, UnaryOp,
-    },
-};
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum SemanticType {
-    Number,
-    Boolean,
-    String,
-    Unknown,
-}
-
-impl SemanticType {
-    fn display_name(self) -> &'static str {
-        match self {
-            SemanticType::Number => "Number",
-            SemanticType::Boolean => "Boolean",
-            SemanticType::String => "String",
-            SemanticType::Unknown => "Unknown",
-        }
-    }
-}
-
-#[derive(Debug, Default)]
-pub struct SemanticAnalyzer {
-    scopes: Vec<HashMap<String, SemanticType>>,
-    errors: Vec<CompilerError>,
-}
-
-impl SemanticAnalyzer {
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    pub fn analyze(&mut self, program: &Program, source: &str) -> Vec<CompilerError> {
-        self.scopes.clear();
-        self.errors.clear();
-        self.push_scope();
-
-        for statement in &program.statements {
-            let _ = self.check_statement(statement, source);
-        }
-
-        self.errors.clone()
-    }
-
-    fn check_statement(&mut self, statement: &Statement, source: &str) -> Option<SemanticType> {
-        match statement {
-            Statement::Let {
-                name,
-                name_span,
-                value,
-                ..
-            } => {
-                if self.is_declared_in_current_scope(name) {
-                    self.push_semantic_error(
-                        *name_span,
-                        source,
-                        format!(
-                            "Variable '{}' redeclared. A variable can only be declared once.",
-                            name
-                        ),
-                    );
-                    return None;
-                }
-
-                let value_type = self
-                    .check_expr(value, source)
-                    .unwrap_or(SemanticType::Unknown);
-                self.current_scope_mut().insert(name.clone(), value_type);
-                Some(value_type)
-            }
-            Statement::Print { value, .. } => self.check_expr(value, source),
-            Statement::Expr { value, .. } => self.check_expr(value, source),
-            Statement::Assign {
-                name,
-                name_span,
-                value,
-                ..
-            } => {
-                let Some(scope_index) = self.find_scope_index(name) else {
-                    self.push_semantic_error(
-                        *name_span,
-                        source,
-                        format!(
-                            "Variable '{}' is assigned before declaration. Declare it with 'let' first.",
-                            name
-                        ),
-                    );
-                    return None;
-                };
-
-                let value_type = self.check_expr(value, source)?;
-                self.scopes[scope_index].insert(name.clone(), value_type);
-                Some(value_type)
-            }
-        }
-    }
-
-    fn check_expr(&mut self, expr: &Expr, source: &str) -> Option<SemanticType> {
-        match expr {
-            Expr::Literal { value, .. } => Some(self.type_of_literal(value)),
-            Expr::DestructiveAssign(assign) => self.check_destructive_assign(assign, source),
-            Expr::Variable { name, span } => {
-                if let Some(var_type) = self.lookup(name) {
-                    Some(var_type)
-                } else {
-                    self.push_semantic_error(
-                        *span,
-                        source,
-                        format!(
-                            "Variable '{}' is used before declaration. Declare it with 'let' first.",
-                            name
-                        ),
-                    );
-                    None
-                }
-            }
-            Expr::Unary(unary) => {
-                let expr_type = self.check_expr(&unary.expr, source)?;
-                match unary.op {
-                    UnaryOp::Neg => {
-                        if expr_type == SemanticType::Unknown {
-                            return Some(SemanticType::Unknown);
-                        }
-
-                        if expr_type == SemanticType::Number {
-                            Some(SemanticType::Number)
-                        } else {
-                            self.push_type_error(
-                                unary.span,
-                                source,
-                                format!(
-                                    "Unary '-' expects Number, but got {}.",
-                                    expr_type.display_name()
-                                ),
-                            );
-                            None
-                        }
-                    }
-                    UnaryOp::Not => {
-                        if expr_type == SemanticType::Unknown {
-                            return Some(SemanticType::Unknown);
-                        }
-
-                        if expr_type == SemanticType::Boolean {
-                            Some(SemanticType::Boolean)
-                        } else {
-                            self.push_type_error(
-                                unary.span,
-                                source,
-                                format!(
-                                    "Unary '!' expects Boolean, but got {}.",
-                                    expr_type.display_name()
-                                ),
-                            );
-                            None
-                        }
-                    }
-                }
-            }
-            Expr::Block(block) => self.check_block_expr(block, source),
-            Expr::LetIn(let_in) => self.check_let_in_expr(let_in, source),
-            Expr::BuiltinCall(call) => {
-                self.check_builtin_call(call.function, &call.args, call.span, source)
-            }
-            Expr::Binary(binary) => {
-                let left_type = self.check_expr(&binary.left, source);
-                let right_type = self.check_expr(&binary.right, source);
-
-                let (Some(left_type), Some(right_type)) = (left_type, right_type) else {
-                    return None;
-                };
-
-                match binary.op {
-                    BinaryOp::Add
-                    | BinaryOp::Sub
-                    | BinaryOp::Mul
-                    | BinaryOp::Div
-                    | BinaryOp::Pow => {
-                        if left_type == SemanticType::Unknown || right_type == SemanticType::Unknown
-                        {
-                            return Some(SemanticType::Unknown);
-                        }
-
-                        if left_type == SemanticType::Number && right_type == SemanticType::Number {
-                            Some(SemanticType::Number)
-                        } else {
-                            let op_name = op_symbol(binary.op.clone());
-                            self.push_type_error(
-                                binary.span,
-                                source,
-                                format!(
-                                    "Operator '{}' expects Number and Number, but got {} and {}.",
-                                    op_name,
-                                    left_type.display_name(),
-                                    right_type.display_name()
-                                ),
-                            );
-                            None
-                        }
-                    }
-                    BinaryOp::Concat => {
-                        if left_type == SemanticType::Unknown || right_type == SemanticType::Unknown
-                        {
-                            return Some(SemanticType::Unknown);
-                        }
-
-                        if is_valid_concat_pair(left_type, right_type) {
-                            Some(SemanticType::String)
-                        } else {
-                            self.push_type_error(
-                                binary.span,
-                                source,
-                                format!(
-                                    "Operator '@' expects (String, String), (String, Number), or (Number, String), but got {} and {}.",
-                                    left_type.display_name(),
-                                    right_type.display_name()
-                                ),
-                            );
-                            None
-                        }
-                    }
-                    BinaryOp::Less
-                    | BinaryOp::Greater
-                    | BinaryOp::LessEqual
-                    | BinaryOp::GreaterEqual => {
-                        if left_type == SemanticType::Unknown || right_type == SemanticType::Unknown
-                        {
-                            return Some(SemanticType::Unknown);
-                        }
-
-                        if left_type == SemanticType::Number && right_type == SemanticType::Number {
-                            Some(SemanticType::Boolean)
-                        } else {
-                            self.push_type_error(
-                                binary.span,
-                                source,
-                                format!(
-                                    "Comparison operator '{}' expects Number and Number, but got {} and {}.",
-                                    op_symbol(binary.op.clone()),
-                                    left_type.display_name(),
-                                    right_type.display_name()
-                                ),
-                            );
-                            None
-                        }
-                    }
-                    BinaryOp::Equal | BinaryOp::NotEqual => {
-                        if left_type == SemanticType::Unknown || right_type == SemanticType::Unknown
-                        {
-                            return Some(SemanticType::Unknown);
-                        }
-
-                        if left_type != right_type {
-                            self.push_type_error(
-                                binary.span,
-                                source,
-                                format!(
-                                    "Operator '{}' expects operands of the same type, but got {} and {}.",
-                                    op_symbol(binary.op.clone()),
-                                    left_type.display_name(),
-                                    right_type.display_name()
-                                ),
-                            );
-                            return None;
-                        }
-
-                        if is_equality_type(left_type) {
-                            Some(SemanticType::Boolean)
-                        } else {
-                            self.push_type_error(
-                                binary.span,
-                                source,
-                                format!(
-                                    "Operator '{}' only supports Number, Boolean, or String operands.",
-                                    op_symbol(binary.op.clone())
-                                ),
-                            );
-                            None
-                        }
-                    }
-                    BinaryOp::And | BinaryOp::Or => {
-                        if left_type == SemanticType::Unknown || right_type == SemanticType::Unknown
-                        {
-                            return Some(SemanticType::Unknown);
-                        }
-
-                        if left_type == SemanticType::Boolean && right_type == SemanticType::Boolean
-                        {
-                            Some(SemanticType::Boolean)
-                        } else {
-                            self.push_type_error(
-                                binary.span,
-                                source,
-                                "logical operator requires Boolean operands".to_string(),
-                            );
-                            None
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    fn check_block_expr(&mut self, block: &BlockExpr, source: &str) -> Option<SemanticType> {
-        self.push_scope();
-
-        let mut last_type: Option<SemanticType> = None;
-        for statement in &block.statements {
-            let stmt_type = self.check_statement(statement, source);
-            if stmt_type.is_some() {
-                last_type = stmt_type;
-            }
-        }
-
-        self.pop_scope();
-
-        if block.statements.is_empty() {
-            Some(SemanticType::Unknown)
-        } else {
-            last_type
-        }
-    }
-
-    fn check_let_in_expr(&mut self, let_in: &LetInExpr, source: &str) -> Option<SemanticType> {
-        self.push_scope();
-
-        for binding in &let_in.bindings {
-            if self.is_declared_in_current_scope(&binding.name) {
-                self.push_semantic_error(
-                    binding.span,
-                    source,
-                    format!("Variable '{}' redeclared in let-in binding.", binding.name),
-                );
-                continue;
-            }
-
-            let value_type = self
-                .check_expr(&binding.value, source)
-                .unwrap_or(SemanticType::Unknown);
-            self.current_scope_mut()
-                .insert(binding.name.clone(), value_type);
-        }
-
-        let body_type = self.check_expr(&let_in.body, source);
-
-        self.pop_scope();
-
-        body_type
-    }
-
-    fn check_builtin_call(
-        &mut self,
-        function: BuiltinFunction,
-        args: &[Expr],
-        span: Span,
-        source: &str,
-    ) -> Option<SemanticType> {
-        match function {
-            BuiltinFunction::Print => {
-                let Some(arg) = args.first() else {
-                    self.push_semantic_error(
-                        span,
-                        source,
-                        "Function 'print' expects 1 argument.".to_string(),
-                    );
-                    return None;
-                };
-                let ty = self.check_expr(arg, source)?;
-                Some(ty)
-            }
-            BuiltinFunction::Sin
-            | BuiltinFunction::Cos
-            | BuiltinFunction::Sqrt
-            | BuiltinFunction::Exp => {
-                let Some(arg) = args.first() else {
-                    self.push_semantic_error(
-                        span,
-                        source,
-                        format!("Function '{}' expects 1 argument.", function.name()),
-                    );
-                    return None;
-                };
-
-                let arg_type = self.check_expr(arg, source)?;
-                if arg_type == SemanticType::Unknown {
-                    return Some(SemanticType::Unknown);
-                }
-
-                if arg_type == SemanticType::Number {
-                    Some(SemanticType::Number)
-                } else {
-                    self.push_type_error(
-                        span,
-                        source,
-                        format!(
-                            "Function '{}' expects Number, but got {}.",
-                            function.name(),
-                            arg_type.display_name()
-                        ),
-                    );
-                    None
-                }
-            }
-            BuiltinFunction::Log => {
-                if args.len() != 2 {
-                    self.push_semantic_error(
-                        span,
-                        source,
-                        "Function 'log' expects 2 arguments.".to_string(),
-                    );
-                    return None;
-                }
-
-                let left_type = self.check_expr(&args[0], source)?;
-                let right_type = self.check_expr(&args[1], source)?;
-                if left_type == SemanticType::Unknown || right_type == SemanticType::Unknown {
-                    return Some(SemanticType::Unknown);
-                }
-
-                if left_type == SemanticType::Number && right_type == SemanticType::Number {
-                    Some(SemanticType::Number)
-                } else {
-                    self.push_type_error(
-                        span,
-                        source,
-                        format!(
-                            "Function 'log' expects (Number, Number), but got {} and {}.",
-                            left_type.display_name(),
-                            right_type.display_name()
-                        ),
-                    );
-                    None
-                }
-            }
-            BuiltinFunction::Rand => {
-                if !args.is_empty() {
-                    self.push_semantic_error(
-                        span,
-                        source,
-                        "Function 'rand' expects 0 arguments.".to_string(),
-                    );
-                    return None;
-                }
-
-                Some(SemanticType::Number)
-            }
-        }
-    }
-
-    fn check_destructive_assign(
-        &mut self,
-        assign: &DestructiveAssignExpr,
-        source: &str,
-    ) -> Option<SemanticType> {
-        let Some(scope_index) = self.find_scope_index(&assign.name) else {
-            self.push_semantic_error(
-                assign.name_span,
-                source,
-                format!(
-                    "Variable '{}' is assigned before declaration. Declare it with 'let' first.",
-                    assign.name
-                ),
-            );
-            return None;
-        };
-
-        let value_type = self.check_expr(&assign.value, source)?;
-        let existing = self.scopes[scope_index]
-            .get(&assign.name)
-            .copied()
-            .unwrap_or(SemanticType::Unknown);
-
-        if existing != SemanticType::Unknown && existing != value_type {
-            self.push_type_error(
-                assign.span,
-                source,
-                format!(
-                    "Destructive assignment ':=' requires the same type. Variable '{}' is {:?} but expression is {:?}.",
-                    assign.name, existing, value_type
-                ),
-            );
-            return None;
-        }
-
-        self.scopes[scope_index].insert(assign.name.clone(), value_type);
-        Some(value_type)
-    }
-
-    fn type_of_literal(&self, literal: &Literal) -> SemanticType {
-        match literal {
-            Literal::Integer(_) | Literal::Float(_) => SemanticType::Number,
-            Literal::Boolean(_) => SemanticType::Boolean,
-            Literal::String(_) => SemanticType::String,
-        }
-    }
-
-    fn push_type_error(&mut self, span: Span, source: &str, message: String) {
-        let (line, column) = offset_to_line_column(source, span.start);
-        self.errors.push(CompilerError::new(
-            ErrorCategory::Type,
-            message,
-            line,
-            column,
-        ));
-    }
-
-    fn push_semantic_error(&mut self, span: Span, source: &str, message: String) {
-        let (line, column) = offset_to_line_column(source, span.start);
-        self.errors.push(CompilerError::new(
-            ErrorCategory::Semantic,
-            message,
-            line,
-            column,
-        ));
-    }
-
-    fn push_scope(&mut self) {
-        self.scopes.push(HashMap::new());
-    }
-
-    fn pop_scope(&mut self) {
-        self.scopes.pop();
-    }
-
-    fn current_scope_mut(&mut self) -> &mut HashMap<String, SemanticType> {
-        self.scopes
-            .last_mut()
-            .expect("at least one scope should be present")
-    }
-
-    fn lookup(&self, name: &str) -> Option<SemanticType> {
-        self.scopes
-            .iter()
-            .rev()
-            .find_map(|scope| scope.get(name).copied())
-    }
-
-    fn find_scope_index(&self, name: &str) -> Option<usize> {
-        self.scopes
-            .iter()
-            .enumerate()
-            .rev()
-            .find_map(|(idx, scope)| scope.contains_key(name).then_some(idx))
-    }
-
-    fn is_declared_in_current_scope(&self, name: &str) -> bool {
-        self.scopes
-            .last()
-            .map(|scope| scope.contains_key(name))
-            .unwrap_or(false)
-    }
-}
-
-fn op_symbol(op: BinaryOp) -> &'static str {
-    match op {
-        BinaryOp::Add => "+",
-        BinaryOp::Pow => "^",
-        BinaryOp::Concat => "@",
-        BinaryOp::Sub => "-",
-        BinaryOp::Mul => "*",
-        BinaryOp::Div => "/",
-        BinaryOp::Equal => "==",
-        BinaryOp::NotEqual => "!=",
-        BinaryOp::Less => "<",
-        BinaryOp::Greater => ">",
-        BinaryOp::LessEqual => "<=",
-        BinaryOp::GreaterEqual => ">=",
-        BinaryOp::And => "&&",
-        BinaryOp::Or => "||",
-    }
-}
-
-fn is_valid_concat_pair(left: SemanticType, right: SemanticType) -> bool {
-    matches!(
-        (left, right),
-        (SemanticType::String, SemanticType::String)
-            | (SemanticType::String, SemanticType::Number)
-            | (SemanticType::Number, SemanticType::String)
-    )
-}
-
-fn is_equality_type(value_type: SemanticType) -> bool {
-    matches!(
-        value_type,
-        SemanticType::Number | SemanticType::Boolean | SemanticType::String
-    )
-}
+pub use analyzer::SemanticAnalyzer;
+pub use helper::SemanticType;
